src/test/java/com/rfacad/rvkybard/KybardJspHelperTest.java
package com.rfacad.rvkybard;

import static org.junit.Assert.*;

import java.io.StringWriter;

import org.junit.Test;

//
//Copyright (c) 2024 Gerald Reno, Jr.
//
//Licensed under the Apache License, Version 2.0 (the "License");
//you may not use this file except in compliance with the License.
//You may obtain a copy of the License at
//
//http://www.apache.org/licenses/LICENSE-2.0
//
public class KybardJspHelperTest
{
    @Test
    public void shouldWriteEndPage()
    {
        StringWriter out = new StringWriter();
        KybardJspHelper h = new KybardJspHelper(out, "", "");
        h.endHtml();
        assertEquals("</body>\r\n</html>\r\n",out.toString());
    }
}
